# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MovieFlix API",
        "description": " API respopnsável por gerenciar os filmes do MovieFlix",
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
            "get": {
                "tags": [
                    " Todos os Filmes"
                ],
                "sumary": "Obter todos os filmes",
                "responses": {
                    "200": {
                        "description": "Retorna uma lista de filmes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "types": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Adicionar um novo filme"
                ],
                "sumary": "Adicionar um novo filme no catalogo",
                "responses": {
                    "201": {
                        "description": "Filme adicionado com sucesso"
                    },
                    "409": {
                        "description": "Esse filme já esta cadastrado"
                    },
                    "500": {
                        "description": "Falha ao cadastrar um filme"
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "tags": [
                    "Atualização dos filmes"
                ],
                "sumary": "atualizar um filme",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filme atualizado com sucesso"
                    },
                    "404": {
                        "description": "Erro ao encontrar o filme"
                    },
                    "500": {
                        "description": "Erro ao atualizar filme"
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Deletar filmes"
                ],
                "sumary": "Apagar um filme",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme que deseja apagar",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filme deletado com sucesso"
                    },
                    "404": {
                        "description": "Filme não encontrado"
                    },
                    "500": {
                        "description": "Erro ao deletar o filme"
                    }
                }
            }
        },
        "/movies/{genreName}": {
            "get": {
                "tags": [
                    "Lista de filmes filtrado pelo gênero"
                ],
                "sumary": "Lista de filmes filtrado pelo gênero",
                "parameters": [
                    {
                        "name": "genreName",
                        "in": "path",
                        "required": true,
                        "description": "Nome do gênero do  filme que você deseja",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "types": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        },
                        "description": "Nova lista de filmes somente com o gênero buscado "
                    },
                    "500": {
                        "description": "Erro ao filtrar gênero"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID do filme"
                    },
                    "title": {
                        "type": "string",
                        "description": "titulo do filme"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Idioma do filme"
                    },
                    "language": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID da idioma"
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do idioma"
                            }
                        }
                    },
                    " genre_id ": {
                        "type": "integer",
                        "description": "Gênero do filme"
                    },
                    "genres": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do gênero"
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do gênero"
                            }
                        }
                    },
                    "realese_data": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de estreia  do filme"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Número de oscar que o filme ganhou"
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": [
                    "title",
                    "language_id",
                    "realese_data",
                    "oscar_count",
                    "genre_id"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Titulo do filme"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "ID do idioma filme"
                    },
                    "realese_data": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de estreia do filme"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Numero de oscar que o filme recebeu"
                    },
                    " genre_id ": {
                        "type": "integer",
                        "description": "ID do gênero do filme"
                    }
                }
            }
        }
    }
}